refactor: move Fred provider to provider/fred sub-package
Use init()-based registration with provider.Register() to avoid import
cycles between provider/ and provider/fred/. Blank import in
cmd/providers_register.go triggers the registration.
